+++ RandomEncounters/Encounter/Flotsam.cs
@@ -1,4 +1,5 @@
 ﻿using UnityEngine;
+using static RandomEncounters.RE_Plugin;
 
 namespace RandomEncounters
 {
@@ -49,62 +50,62 @@ namespace RandomEncounters
         //"317 crate of tobacco brown",
         //"319 crate of tobacco blue"
 
-        private static readonly int[] cargos = { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 13, 14, 15, 16, 17, 18, 19, 24, 25, 26, 27 };
-        private static readonly int[] consumables = { 104, 108, 131, 132 };       
-        private static readonly int[] bottles = { 55, 56, 57, 58, 59 };
-        private static readonly int[] tobaccos = { 311, 313, 315, 319 };
+        private static readonly int[] _cargos = { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 13, 14, 15, 16, 17, 18, 19, 24, 25, 26, 27 };
+        private static readonly int[] _consumables = { 104, 108, 131, 132 };
+        private static readonly int[] _bottles = { 55, 56, 57, 58, 59 };
+        private static readonly int[] _tobaccoCrates = { 311, 313, 315, 319 };
 
-        public static void Spawn(Vector3 spawnPoint)
+        private const int MAX_CARGO_TYPES = 4;
+        private const int MAX_CARGOS = 3;
+        private const int MAX_CONSUME_TPYES = 2;
+
+        internal static void Spawn(Vector3 spawnPoint)
         {
-            var maxCargosTypes = 4;
-            var maxCargos = 3;
-            var maxConsumTypes = 2;
-
-            for (int i = 0; i < Random.Range(1, maxCargosTypes); i++)
+            for (int i = 0; i < Random.Range(1, MAX_CARGO_TYPES); i++)
             {
-                var choice = Random.Range(0, cargos.Length - 1);
+                var choice = Random.Range(0, _cargos.Length - 1);
 
-                for (int j = 0; j < Random.Range(1, maxCargos); j++)
+                for (int j = 0; j < Random.Range(1, MAX_CARGOS); j++)
                 {
-                    Plugin.logger.LogDebug($"Choice: {cargos[choice]}");
-                    var prefabGO = PrefabsDirectory.instance.directory[cargos[choice]];
+                    LogDebug($"Choice: {_cargos[choice]}");
+                    var prefabGO = PrefabsDirectory.instance.directory[_cargos[choice]];
                     var amount = (float)System.Math.Round((decimal)Random.Range(0, prefabGO.GetComponent<ShipItem>().amount));
                     SpawnItem(spawnPoint, prefabGO, amount);
-                }                    
+                }
             }
 
-            for (int i = 0; i < Random.Range(1, maxConsumTypes); i++)
+            for (int i = 0; i < Random.Range(1, MAX_CONSUME_TPYES); i++)
             {
-                var choice = Random.Range(0, consumables.Length - 1);
-                Plugin.logger.LogDebug($"Choice: {consumables[choice]}");
-                var prefabGO = PrefabsDirectory.instance.directory[consumables[choice]];
+                var choice = Random.Range(0, _consumables.Length - 1);
+                LogDebug($"Choice: {_consumables[choice]}");
+                var prefabGO = PrefabsDirectory.instance.directory[_consumables[choice]];
                 var amount = (float)System.Math.Round((decimal)Random.Range(0, prefabGO.GetComponent<ShipItem>().amount));
-                SpawnItem(spawnPoint, prefabGO, amount);                
+                SpawnItem(spawnPoint, prefabGO, amount);
             }
 
             for (int i = 0; i < Random.Range(5, 10); i++)
             {
-                var choice = Random.Range(0, bottles.Length - 1);
-                Plugin.logger.LogDebug($"Choice: {bottles[choice]}");
-                var prefabGO = PrefabsDirectory.instance.directory[bottles[choice]];
+                var choice = Random.Range(0, _bottles.Length - 1);
+                LogDebug($"Choice: {_bottles[choice]}");
+                var prefabGO = PrefabsDirectory.instance.directory[_bottles[choice]];
                 var amount = 0f;
                 SpawnItem(spawnPoint, prefabGO, amount);
             }
 
-            var tobaccoChoice = Random.Range(0, tobaccos.Length - 1);
-            Plugin.logger.LogDebug($"Choice: {tobaccos[tobaccoChoice]}");
-            var tobaccoPrefabGO = PrefabsDirectory.instance.directory[tobaccos[tobaccoChoice]];
+            var tobaccoChoice = Random.Range(0, _tobaccoCrates.Length - 1);
+            LogDebug($"Choice: {_tobaccoCrates[tobaccoChoice]}");
+            var tobaccoPrefabGO = PrefabsDirectory.instance.directory[_tobaccoCrates[tobaccoChoice]];
             var tobaccoAmount = (float)System.Math.Round((decimal)Random.Range(0, tobaccoPrefabGO.GetComponent<ShipItem>().amount));
             SpawnItem(spawnPoint, tobaccoPrefabGO, tobaccoAmount);
 
-            SpawnItem(spawnPoint, AssetLoader.hull, 1f, true);
-            SpawnItem(spawnPoint, AssetLoader.mast, 1f, true);
-            SpawnItem(spawnPoint, AssetLoader.bowsprit, 1f, true);
+            SpawnItem(spawnPoint, AssetLoader.Hull, 1f, true);
+            SpawnItem(spawnPoint, AssetLoader.Mast, 1f, true);
+            SpawnItem(spawnPoint, AssetLoader.Bowsprit, 1f, true);
         }
 
-        public static void SpawnItem(Vector3 spawnPoint, GameObject prefabGO, float amount, bool wreckage = false)
+        internal static void SpawnItem(Vector3 spawnPoint, GameObject prefabGO, float amount, bool wreckage = false)
         {
-            GameObject obj = Object.Instantiate(prefabGO, spawnPoint, Quaternion.Euler(Random.Range(0, 360), Random.Range(0, 360), Random.Range(0, 360)));
+            var obj = Object.Instantiate(prefabGO, spawnPoint, Quaternion.Euler(Random.Range(0, 360), Random.Range(0, 360), Random.Range(0, 360)));
             obj.GetComponent<ShipItem>().sold = true;
             obj.GetComponent<SaveablePrefab>().RegisterToSave();
             if ((bool)obj.GetComponent<Good>())
@@ -115,8 +116,8 @@ namespace RandomEncounters
             {
                 obj.GetComponent<ShipItem>().unclickable = true;
                 obj.transform.parent = Refs.shiftingWorld;
-            }                
-            Plugin.logger.LogDebug($"Prefab {prefabGO.name} spawned");
+            }
+            LogDebug($"Prefab {prefabGO.name} spawned");
         }
     }
 }