+++ RandomEncounters/EncounterGenerator.cs
@@ -1,41 +1,75 @@
 ﻿using System.Collections;
-using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
+using static RandomEncounters.RE_Plugin;
+using static RandomEncounters.Configs;
 
 namespace RandomEncounters
 {
     internal class EncounterGenerator : MonoBehaviour
     {
-        public static EncounterGenerator instance;
-        public static List<Transform> whaleSpawns;
-        private bool moveSeagulls;
+        public static EncounterGenerator Instance { get; private set; }
+
+        private bool _moveSeagulls;
+
+        private const float MIN_DISTANCE = 1000F;
+        private const int ENCOUNTER_TIME_RANGE = 300;
 
         public void Awake()
         {
-            instance = this;
-            whaleSpawns = new List<Transform>();
-            StartCoroutine(GenerateEncounters());
-        }
-
-        private IEnumerator GenerateEncounters()
-        {
-            while (true)
+            if (Instance != null && Instance != this)
             {
-                yield return new WaitForSeconds(Random.Range(Plugin.generateEncounterMinTime.Value, Plugin.generateEncounterMinTime.Value + 300));
-                if (GameState.currentBoat != null)
-                    Generate();
+                Destroy(gameObject);
+                return;
             }
+            Instance = this;
+            
+            StartCoroutine(ScheduleEncounter());
         }
 
-        internal void Generate()
+        public void Update()
         {
-            Plugin.logger.LogDebug($"Distance to land {GameState.distanceToLand}");
+            SeaLifeMod.CheckWhaleDistance();
 
-            if (!GameState.playing || GameState.sleeping || GameState.distanceToLand <= 1000f) return;
+            /*
+            // for testing
+            if (Input.GetKeyDown(KeyCode.P))
+            {
+                Generate();
+            }
+            */
+        }
+
+        private IEnumerator ScheduleEncounter()
+        {
+            var maxTime = generateEncounterMinTime.Value + ENCOUNTER_TIME_RANGE;
+            var timeToNextEncounter = Random.Range(generateEncounterMinTime.Value, maxTime);
+            yield return new WaitForSeconds(timeToNextEncounter);
+            if (GameState.currentBoat != null)
+                Generate();
+
+            StartCoroutine(ScheduleEncounter());
+        }
+
+        private void Generate()
+        {
+            if (!GameState.playing)
+                return;
+
+            if (GameState.distanceToLand <= MIN_DISTANCE)
+            {
+                LogDebug($"Too close to land (distance: {GameState.distanceToLand}), skipping encounter generation.");
+                return;
+            }
+
+            if (GameState.sleeping)
+            {
+                LogDebug("Player sleeping, skipping encounter generation.");
+                return;
+            }            
 
             var roll = Random.Range(1, 100);
-            Plugin.logger.LogDebug($"Roll: {roll}");
+            LogDebug($"Roll: {roll}");
 
             switch (roll)
             {
@@ -44,10 +78,10 @@ namespace RandomEncounters
                     break;
                 case int n when n > 10 && n <= 15:
                     GenerateFlotsam();
-                    StartCoroutine(GenerateWhale());
+                    StartCoroutine(GenerateWhales());
                     break;
                 case int n when n > 15 && n <= 40:
-                    StartCoroutine(GenerateWhale());
+                    StartCoroutine(GenerateWhales());
                     break;
                 case int n when n > 40 && n <= 45:
                     StartCoroutine(GenerateDenseFog());
@@ -61,48 +95,53 @@ namespace RandomEncounters
             }
         }
 
-        internal static void GenerateFlotsam()
+        private static void GenerateFlotsam()
         {
-            if (!Plugin.enableFlotsam.Value) return;
+            if (!enableFlotsam.Value) 
+                return;
 
             var spawnPoint = GameState.currentBoat.position + GameState.currentBoat.right * 100f + GameState.currentBoat.forward * Random.Range(-30, 30);
             Flotsam.Spawn(spawnPoint);
         }
 
-        internal IEnumerator GenerateWhale()
+        #region whales
+
+        private IEnumerator GenerateWhales()
         {
-            if (!Plugin.controlSeaLifeMod.Value || Plugin.seaLifeModInstance == null) 
+            if (!controlSeaLifeMod.Value || SeaLifeModPluginInstance == null) 
                 yield break;
 
-            for (int i = 0; i < Random.Range(1, 3); i++)
+            var boatPosition = GameState.currentBoat.position;            
+
+            for (int i = 0; i < Random.Range(2, 5); i++)
             {
-                Plugin.logger.LogDebug($"Spawning Finwhale");
-                var seaLifespawnPoint = GameState.currentBoat.position + new Vector3(Random.Range(-200, 200), -8, Random.Range(-200, 200));
-                SeaLifeMod.spawnWhale(Plugin.seaLifeModInstance, seaLifespawnPoint);
-                yield return new WaitForSeconds(0.5f);
+                var randomOffset = new Vector3(Random.Range(-200, 200), -8, Random.Range(-200, 200));
+                yield return new WaitForSeconds(2f);                
+                SeaLifeMod.SpawnWhale(i, boatPosition + randomOffset);
             }
             yield return new WaitForSeconds(2f);
-            whaleSpawns = Refs.shiftingWorld.GetComponentsInChildren<Transform>().Where(t => t.name == "FinWhalePrefab(Clone)").ToList();
-            var whaleTransform = whaleSpawns.FirstOrDefault();
-            var finWhaleAI = whaleTransform.gameObject.GetComponent("FinWhaleAI");
-            SeaLifeMod.triggerRandomAnimation(finWhaleAI);            
-        }
+            SeaLifeMod.TriggerEntranceAnimation();                      
+        }        
 
-        internal IEnumerator GenerateDenseFog()
+        #endregion
+
+        #region dense fog
+
+        private IEnumerator GenerateDenseFog()
         {
-            if (!Plugin.enableDenseFog.Value ||
-                DenseFog.running ||
+            if (!enableDenseFog.Value ||
+                DenseFog.IsRunning ||
                 WeatherStorms.instance.InvokePrivateMethod<float>("GetNormalizedDistance") < 0.75f)
                 yield break;
 
             DenseFog.Spawn();
-            for (int i = 0; i < 4000; i++) 
+            for (int i = 0; i < 4000; i++)
             {
-                foreach (var audioSource in DenseFog.waveAudioSources.Keys)
+                foreach (var audioSource in DenseFog.WaveAudioSources.Keys)
                 {
-                    audioSource.volume = Mathf.Lerp(DenseFog.waveAudioSources[audioSource], 0f, i / 1000f);
+                    audioSource.volume = Mathf.Lerp(DenseFog.WaveAudioSources[audioSource], 0f, i / 1000f);
                 }
-                DenseFog.windAudioSource.source.volume = Mathf.Lerp(DenseFog.windAudioSource.origVolume, 0.0001f, i / 1000f);
+                DenseFog.WindAudioSource.source.volume = Mathf.Lerp(DenseFog.WindAudioSource.origVolume, 0.0001f, i / 1000f);
                 yield return new WaitForSeconds(0.001f);
             }
 
@@ -113,31 +152,35 @@ namespace RandomEncounters
                     GameState.currentBoat.right * (200f + Random.Range(20f, 60f) * i) +
                     GameState.currentBoat.forward * Random.Range(-200, 200);
 
-                Flotsam.SpawnItem(spawnPoint, AssetLoader.hull, 1f, true);
+                Flotsam.SpawnItem(spawnPoint, AssetLoader.Hull, 1f, true);
                 yield return new WaitForSeconds(1f);
-                Flotsam.SpawnItem(spawnPoint, AssetLoader.mast, 1f, true);
+                Flotsam.SpawnItem(spawnPoint, AssetLoader.Mast, 1f, true);
                 yield return new WaitForSeconds(1f);
-                Flotsam.SpawnItem(spawnPoint, AssetLoader.bowsprit, 1f, true);
+                Flotsam.SpawnItem(spawnPoint, AssetLoader.Bowsprit, 1f, true);
                 yield return new WaitForSeconds(1f);
             }
 
-            yield return new WaitForSeconds(Plugin.fogDuration.Value);
+            yield return new WaitForSeconds(fogDuration.Value);
 
             DenseFog.ClearFog();
             for (int i = 0; i < 1000; i++)
             {
-                foreach (var audioSource in DenseFog.waveAudioSources.Keys)
+                foreach (var audioSource in DenseFog.WaveAudioSources.Keys)
                 {
-                    audioSource.volume = Mathf.Lerp(0f, DenseFog.waveAudioSources[audioSource], i / 1000f);
+                    audioSource.volume = Mathf.Lerp(0f, DenseFog.WaveAudioSources[audioSource], i / 1000f);
                 }
-                DenseFog.windAudioSource.source.volume = Mathf.Lerp(0.0001f, DenseFog.windAudioSource.origVolume,  i / 1000f);
+                DenseFog.WindAudioSource.source.volume = Mathf.Lerp(0.0001f, DenseFog.WindAudioSource.origVolume,  i / 1000f);
                 yield return new WaitForSeconds(0.001f);
             }
         }
 
-        internal IEnumerator GenerateFishingBonanza()
+        #endregion
+
+        #region fishing bonanza
+
+        private IEnumerator GenerateFishingBonanza()
         {
-            if (!Plugin.enableFishingBonanza.Value ||
+            if (!enableFishingBonanza.Value ||
             GameState.currentBoat == null ||
             WeatherStorms.instance.InvokePrivateMethod<float>("GetNormalizedDistance") < 0.75f)
             {
@@ -147,7 +190,7 @@ namespace RandomEncounters
             var seagullsGO = Refs.islands[3].GetComponentsInChildren<Transform>().FirstOrDefault(t => t.name == "seagulls")?.gameObject;
             if (seagullsGO == null)
             {
-                Plugin.logger.LogDebug("No seagulls found");
+                LogDebug("No seagulls found");
                 yield break;
             }
 
@@ -160,30 +203,34 @@ namespace RandomEncounters
             var emission = seagullsPS.emission;
             if (!emission.enabled) emission.enabled = true;
 
-            Plugin.logger.LogDebug("Starting fishing bonanza");
-            FishingBonanza.bonanzaActive = true;
-            moveSeagulls = true;
+            LogDebug("Starting fishing bonanza");
+            FishingBonanza.IsBonanzaActive = true;
+            _moveSeagulls = true;
             StartCoroutine(MoveSegulls(seagulls.transform, GameState.currentBoat));
-            yield return new WaitForSeconds(Plugin.fishingBonanzaDuration.Value);            
+            yield return new WaitForSeconds(fishingBonanzaDuration.Value);            
 
-            Plugin.logger.LogDebug("Stopping fishing bonanza");
-            FishingBonanza.bonanzaActive = false;
-            moveSeagulls = false;
+            LogDebug("Stopping fishing bonanza");
+            FishingBonanza.IsBonanzaActive = false;
+            _moveSeagulls = false;
             Destroy(seagulls);
         }
 
         private IEnumerator MoveSegulls(Transform seagulls, Transform boat)
         {
-            while (moveSeagulls)
+            while (_moveSeagulls)
             {
                 seagulls.position = boat.position + boat.up * 60f;
                 yield return null;
-            }            
+            }
         }
 
-        internal IEnumerator GenerateIntenseStorm()
+        #endregion
+
+        #region intense storm
+
+        private IEnumerator GenerateIntenseStorm()
         {
-            if (!Plugin.enableIntenseStorm.Value)
+            if (!enableIntenseStorm.Value)
                 yield break;
 
             var weatherStorms = WeatherStorms.instance;
@@ -201,10 +248,10 @@ namespace RandomEncounters
             targetRegion.stormWeather.particles.rainDensity = 70f;
 
             var stormDist = Vector3.Distance(Camera.main.transform.position, storm.transform.position);
-            Vector3 vector = Camera.main.transform.position - storm.transform.position;
+            var vector = Camera.main.transform.position - storm.transform.position;
             vector.y = 0f;
 
-            Plugin.logger.LogDebug($"{storm.name} approaching");
+            LogDebug($"{storm.name} approaching");
             while (stormDist > 1500f)
             {
                 vector = Camera.main.transform.position - storm.transform.position;                
@@ -216,17 +263,17 @@ namespace RandomEncounters
                 stormDist = Vector3.Distance(Camera.main.transform.position, storm.transform.position);
             }
 
-            Plugin.logger.LogDebug($"{storm.name} arrived");
+            LogDebug($"{storm.name} arrived");
             IntenseStorm.oceanUpdaterCrest.inertiaWindScale = 0.22f;
             IntenseStorm.oceanUpdaterCrest.SetPrivateField("windSpeedMult", 5f);
             IntenseStorm.oceanUpdaterCrest.SetPrivateField("smallWavesMult", 0.4f);
-            for (int i = 0; i < Plugin.intenseStormDuration.Value; i++)
+            for (int i = 0; i < intenseStormDuration.Value; i++)
             {
                 Wind.currentBaseWind = vector * 50f;
                 yield return new WaitForSeconds(1f);
             }
 
-            Plugin.logger.LogDebug($"{storm.name} dying down");
+            LogDebug($"{storm.name} dying down");
             lightning.SetPrivateField("lightningInterval", origLightningInterval);
             Weather.instance.currentRegion.stormWeather.particles.rainDensity = origRainDensity;
             IntenseStorm.oceanUpdaterCrest.inertiaWindScale = origInertiaWindScale;
@@ -234,14 +281,6 @@ namespace RandomEncounters
             IntenseStorm.oceanUpdaterCrest.SetPrivateField("smallWavesMult", origSmallWavesMult);            
         }
 
-        /*
-        // for testing
-        void Update()
-        {
-            if (Input.GetKeyDown(KeyCode.P))
-            {
-                Generate();
-            }
-        }
-        */
-    }}
+        #endregion
+    }
+}