+++ RandomEncounters/Utilities/AssetLoader.cs
@@ -1,33 +1,36 @@
-﻿using System.IO;
+﻿using System.Collections;
+using System.IO;
+using System.Linq;
 using UnityEngine;
+using static RandomEncounters.RE_Plugin;
 
 namespace RandomEncounters
 {
     internal class AssetLoader
     {
-        internal static GameObject hull;
-        internal static GameObject mast;        
-        internal static GameObject bowsprit;
+        public static GameObject Hull { get; private set; }
+        public static GameObject Mast { get; private set; }
+        public static GameObject Bowsprit { get; private set; }
 
-        internal static void LoadFlotsam()
+        internal static IEnumerator LoadAssetBundle()
         {
-            var bundlePath = Path.Combine(Path.GetDirectoryName(Plugin.instance.Info.Location), "Assets", "wreckage_bundle");
-            var assetBundle = LoadAssetBundle(bundlePath);
-            hull = assetBundle.LoadAsset<GameObject>("hull");
-            mast = assetBundle.LoadAsset<GameObject>("mast");
-            bowsprit = assetBundle.LoadAsset<GameObject>("bowsprit");
-        }
+            LogDebug("Loading bundle");
+            var bundlePath = Path.Combine(Path.GetDirectoryName(Instance.Info.Location), "Assets", "wreckage_bundle");            
+            var assetBundleRequest = AssetBundle.LoadFromFileAsync(bundlePath);
+            yield return assetBundleRequest;
 
-        public static AssetBundle LoadAssetBundle(string path)
-        {
-            
-            var bundle = AssetBundle.LoadFromFile(path);
-            if (bundle == null)
-            {
-                Plugin.logger.LogError($"Failed to load {path}");
-                return null;
-            }
-            return bundle;
+            var assetBundle = assetBundleRequest.assetBundle;
+            if (assetBundle == null)            
+                LogError($"Failed to load {bundlePath}");
+            var request = assetBundle.LoadAllAssetsAsync();
+            yield return request;
+
+            Hull = request.allAssets.FirstOrDefault(a => a.name == "hull") as GameObject;
+            Mast = request.allAssets.FirstOrDefault(a => a.name == "mast") as GameObject;
+            Bowsprit = request.allAssets.FirstOrDefault(a => a.name == "bowsprit") as GameObject;
+
+            if(Hull == null || Mast == null || Bowsprit == null)            
+                LogError("Failed to load all assets from the bundle.");            
         }
     }
 }